requirements.txt
asgiref==3.9.0
Django==5.2.9
django-allauth==65.13.1
django-cors-headers==4.0.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
Pillow==10.0.0
psycopg2-binary==2.9.6
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.14.1
